+++ apps/web/src/components/organisms/MapComponent/index.tsx
@@ -3,10 +3,10 @@
 import type { ReactElement } from "react"
 import "mapbox-gl/dist/mapbox-gl.css"
 import { useCallback, useEffect, useState } from "react"
-import { DebugPanel } from "@/components/atoms/DebugPanel"
-import { SoundPinMarkers } from "@/components/molecules/SoundPinMarkers"
 import { UserMarker } from "@/components/molecules/UserMarker"
+import { DebugPanel } from "@/components/organisms/DebugPanel"
 import { PinAudioPlayer } from "@/components/organisms/PinAudioPlayer"
+import { SoundPinMarkers } from "@/components/organisms/SoundPinMarkers"
 import { useMapRealtime } from "@/hooks/useRealtime"
 import type { MapBounds, SoundPin } from "@/store/useSoundPinStore"
 import { useMapComponent } from "./hooks"
